improve startup speed and logging feedback
Switched runtime to tokio for ease of usage and even some speed improvements!
A logging library has been introduced to better format output and to
possibly have logging to a file in the future.
If a program can't be executed the user now gets a notification,
informing them of the reason why.